requirements.txt
asgiref==3.3.1
Django==3.1.5
django-google-maps==0.12.2
djangorestframework==3.12.2
geographiclib==1.50
geopy==2.1.0
pytz==2020.5
sqlparse==0.4.1
